requirements.txt
streamlit
mysql-connector-python
bcrypt
pandas
python-dotenv
